# Refactoring CTF functions