--- a/pyproject.toml
+++ b/pyproject.toml
@@ -19,7 +19,7 @@ name = "torch-fourier-filter"
 dynamic = ["version"]
 description = "Filters for image and volumes in pyTorch"
 readme = "README.md"
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 license = { text = "BSD-3-Clause" }
 authors = [{ name = "Josh Dickerson", email = "jdickerson@berkeley.edu" }]
 # https://pypi.org/classifiers/