This is a refactor of the CTF functions, with the following goals:

- Make the calculation of the actual CTF independent of the dimensions
- Make it easier to extract the phase abberation (for equiphase averaging)
- Make the calculation of astigmatism easier to understand (for me at least, with an idea for a different parameterization, explained below)

Nothing should have changed about the input/output of calculate_ctf_xd, at least according to the tests (I've added some for astig/b-factor)

There is a new function called `calculate_phase_aberration` that takes care of the actual calculation and handles the conversions from the various units to angstroms. I takes a squared frequencies grid as an input, which can be an arbitrary tensor. The function, however, requires that all inputs are tensors that are broadcastable with the frequency grid in a way that makes sense. The `calculate_ctf_xd` functions now only take care of setting up the frequency grid and make sure all parameters are broadcastable.

I changed the way the astigmatism is calculated. In the 2D function the defocus tensor has no longer singleton dimensions for the frequency but is instead passed with the effective defocus value at each point in the 2D grid. This is calculated by converting the astig amount/angle to a vector and the defocus offset is calculated as the squared dotproduct of this vector with the frequency vector normalized to a unitvector. The ideas behind this are:

- In my head a dotproduct is faster than calculating the azimuth angle (have not bench-marked this)
- I have this vague idea that the adam optimizer would work better if it was optimizing the astigmatism as a vector. The problem with the astigmatism amount + angle definition is that the angle is not really defined when the astigmatism is 0, which must generate weird gradients.

Some other things:

- I think for the B-factor the frequencies should be squared, but before it used n4, which was the frequency**4 (?)

Happy to discuss if somebody things differently about these things and only take parts of this PR (such as the tests)