@jdickerson95 2025-07-10
I like the changes, I was looking at it yesterday and thinking about separating out the phase aberration.
One point is the definition of phase aberration. To me, the k5 term is the amplitude term and not part of a phase aberration. I would maybe separate this out into a separate function.

With respect to the B factor. Looks like it should be squared to me. This was copied form libtilt and looks like it follows RELIONs format to me, so @alisterburt may have more insight on that.