@jojoelfe 2025-07-10
Is the B factor used anywhere? Otherwise, I can just pull it.

Get Outlook for Android<https://aka.ms/AAb9ysg>
________________________________
From: Alister Burt ***@***.***>
Sent: Thursday, July 10, 2025 4:43:11 PM
To: teamtomo/torch-fourier-filter ***@***.***>
Cc: Johannes Elferich ***@***.***>; Author ***@***.***>
Subject: Re: [teamtomo/torch-fourier-filter] Refactoring CTF functions (PR #12)


@alisterburt commented on this pull request.

loooove to see this - big fan

few minor nits, responded inline re: n4 usage - please feel free to address of ignore as you see fit

I agree the implementation is easier to follow and love that :)

________________________________

In src/torch_fourier_filter/ctf.py<https://github.com/teamtomo/torch-fourier-filter/pull/12#discussion_r2198658340>:

> +    phase_shift: torch.Tensor,
+    amplitude_contrast: torch.Tensor,


nit: no units when rest of func has units

________________________________

In src/torch_fourier_filter/ctf.py<https://github.com/teamtomo/torch-fourier-filter/pull/12#discussion_r2198678639>:

> +    astigmatism_vector = torch.stack(
+        [
+            torch.sin(torch.deg2rad(astigmatism_angle)),
+            torch.cos(torch.deg2rad(astigmatism_angle)),
+        ],
+        dim=-1,
+    ) * torch.sqrt(einops.rearrange(astigmatism, "... -> ... 1"))


nit: slightly unclear what's going on here because of inlined reshapes

something like this maybe?

⬇️ Suggested change

-    astigmatism_vector = torch.stack(
-        [
-            torch.sin(torch.deg2rad(astigmatism_angle)),
-            torch.cos(torch.deg2rad(astigmatism_angle)),
-        ],
-        dim=-1,
-    ) * torch.sqrt(einops.rearrange(astigmatism, "... -> ... 1"))
+    sin_theta = torch.sin(torch.deg2rad(astigmatism_angle))
+    cos_theta = torch.cos(torch.deg2rad(astigmatism_angle))
+    unit_astigmatism_vector_yx = einops.rearrange([sin_theta, cos_theta], "yx ... -> ... yx")
+    astigmatism = einops.rearrange(astigmatism, "... -> ... 1")
+    astigmatism_vector = astigmatism * unit_astigmatism_vector_yx


________________________________

In src/torch_fourier_filter/ctf.py<https://github.com/teamtomo/torch-fourier-filter/pull/12#discussion_r2198686703>:

> +    if torch.any(b_factor > 0):
+        k4 = -b_factor / 4
         idx = (k4 < 0).squeeze(dim=(-2,-1))
-        ctf[idx] *= torch.exp(k4[idx] * n4[idx])
+        ctf[idx] *= torch.exp(k4[idx] * fft_freq_grid_squared[idx])


I honestly never loved having B factor in the implementation and we can definitely consider keeping the envelopes separate - I'm not sure why the square of the norms was used here either, I followed a Python implementation that originated somewhere in RELION...

—
Reply to this email directly, view it on GitHub<https://github.com/teamtomo/torch-fourier-filter/pull/12#pullrequestreview-3007240859>, or unsubscribe<https://github.com/notifications/unsubscribe-auth/ABOMUD6RWAY24QXGR5OKJKL3H3F57AVCNFSM6AAAAACBH4MMRSVHI2DSMVQWIX3LMV43YUDVNRWFEZLROVSXG5CSMV3GSZLXHMZTAMBXGI2DAOBVHE>.
You are receiving this because you authored the thread.Message ID: ***@***.***>