@jdickerson95 2025-07-10
There's a b_envelope in envelopes.py anyway. So it can be used there if necessary.