@McHaillet 2025-07-11
> I guess a bigger question is do we (TeamTomo) want to include envelope functions in the CTF calculation, or should they exist somewhere else?

Hmm, I'd say it makes sense to have a envelopes in the CTF, as it represents the whole contrast transfer of the microscope