@jojoelfe 2025-07-11
> @alisterburt We (@jdickerson95 and I) are for 2DTM for the envelope function. We have a script to find the "optimal" B-factor for a given reference model which varies the CTF B-factor rather than globally applied B-factors during a simulation. Cuts down on computational cost since the simulated model remains static at each iteration.

Did this give reasonable results? I just threw together a quick test on how the b-factor envelope looked before and after this change:

```python
from torch_fourier_filter.ctf import calculate_ctf_1d
import matplotlib.pyplot as plt
plt.plot(calculate_ctf_1d(
    pixel_size=1.0,
    defocus=2.0,
    voltage=300.0,
    spherical_aberration=2.0,
    amplitude_contrast=0.1,
    b_factor=100.0,
    phase_shift=0.0,
    oversampling_factor=1,
    n_samples=512
    ))
plt.show()
```

# This PR
```bash
uv run python plot.py
```
<img width="569" height="416" alt="image" src="https://github.com/user-attachments/assets/0c5401cc-36b9-4fc1-afe6-ce7b5c6686e6" />

# torch-fourier-filter 0.2.5
```bash
uv run --with torch-fourier-filter==0.2.5 --with matplotlib --python 3.10  Documents/torch-fourier-filter/plot.py 
```

<img width="569" height="416" alt="image" src="https://github.com/user-attachments/assets/9e18cb0a-1039-4cdf-916f-b84cf7c22580" />

I guess in the end it doesn't really matter because the absolute value of the B factor has so little meaning. Ours were off by a factor of 4 for a long time, because the cistem simulator did not divide it by 4 in the exponent. (and maybe the **4 shape is better?)