@jdickerson95 2025-07-11
#

> Did this give reasonable results? I just threw together a quick test on how the b-factor envelope looked before and after this change:

The optimal B was slightly higher than I expected in some cases, but nothing crazy. Looks like that makes sense based on your plots. 

> Hmm, I'd say it makes sense to have a envelopes in the CTF, as it represents the whole contrast transfer of the microscope

I think this could get messy. In my mind, the B-factor is more capturing unmodelled parts of the envelope, usually not to do with the microscope (i.e. particle flexibility, radiation damage...). If we were putting all envelopes in the CTF, then the energy spread would need to be defined for the Cc envelope, the source semi-angle for the Cs envelope, stage vibration and drift. These get a little complicated and are simple to calculate separately and apply.

The B-factor is simple and acts as a bit of a catch-all, so I'm less opposed to it being in the ctf. I would probably remove it, but don't mind either way.