@jojoelfe 2025-07-12
I think I addressed most of the nitpicks, but decided to checkup on the performance thing with a script from claude:

https://gist.github.com/jojoelfe/eca13869d798fc53217986c8b593f3bf

The results are quite interesting:

# Before
```
uv run --with "git+https://github.com/teamtomo/torch-fourier-filter.git" --python 3.10 https://gist.githubusercontent.com/jojoelfe/eca13869d
798fc53217986c8b593f3bf/raw/2491b06580ed4fa0c675e9617feac43a7b9b65d9/benchmark_ctf_2d.py
      Built torch-fourier-filter @ git+https://github.com/teamtomo/torch-fourier-filter.git@b90ac87dc5d7519b3a33f89617d4a52637e9c5f2
Installed 37 packages in 26.18s
PyTorch version: 2.7.1+cu126
CUDA available: True
CUDA device: NVIDIA GeForce RTX 3090
CUDA memory: 23.6 GB

================================================================================
CTF 2D Benchmark Results
================================================================================
Device   Image Size   Batch    Time (ms)    Memory (MB)  Throughput     
--------------------------------------------------------------------------------
cpu      64x64           1        0.31         0.02         3214.52        
cpu      64x64           8        0.44         0.12         18354.81       
cpu      64x64           32       0.74         0.50         43141.57       
cpu      128x128          1        0.47         0.06         2116.62        
cpu      128x128          8        1.18         0.50         6797.62        
cpu      128x128          32       1.98         2.00         16195.01       
cpu      256x256          1        1.08         0.25         924.20         
cpu      256x256          8        2.16         2.00         3705.34        
cpu      256x256          32       8.58         8.00         3730.93        
cpu      512x512          1        1.51         1.00         660.20         
cpu      512x512          8        8.90         8.00         898.78         
cpu      512x512          32       63.09        32.00        507.25         
cuda     64x64           1        0.44         0.22         2254.53        
cuda     64x64           8        0.42         1.65         19167.96       
cuda     64x64           32       0.42         6.53         76317.23       
cuda     128x128          1        0.41         1.28         2422.97        
cuda     128x128          8        0.41         6.54         19557.07       
cuda     128x128          32       0.43         26.05        74446.31       
cuda     256x256          1        0.41         5.05         2451.19        
cuda     256x256          8        0.43         26.06        18711.87       
cuda     256x256          32       0.78         104.07       40969.50       
cuda     512x512          1        0.42         20.07        2354.24        
cuda     512x512          8        0.79         104.08       10100.44       
cuda     512x512          32       2.07         416.09       15447.88       

================================================================================
Performance Analysis
================================================================================
Best GPU performance: 76317.23 images/sec (64x64, batch=32)
Best CPU performance: 43141.57 images/sec (64x64, batch=32)
GPU speedup: 1.77x

Benchmark completed!
```

# After
```
uv run --with "git+https://github.com/teamtomo/torch-fourier-filter.git@pull/12/head" --python 3.10 https://gist.githubusercontent.com/jojoe
lfe/eca13869d798fc53217986c8b593f3bf/raw/2491b06580ed4fa0c675e9617feac43a7b9b65d9/benchmark_ctf_2d.py
Installed 37 packages in 25.97s
PyTorch version: 2.7.1+cu126
CUDA available: True
CUDA device: NVIDIA GeForce RTX 3090
CUDA memory: 23.6 GB

================================================================================
CTF 2D Benchmark Results
================================================================================
Device   Image Size   Batch    Time (ms)    Memory (MB)  Throughput     
--------------------------------------------------------------------------------
cpu      64x64           1        0.65         0.02         1534.58        
cpu      64x64           8        2.83         0.12         2825.09        
cpu      64x64           32       11.06        0.50         2892.69        
cpu      128x128          1        1.59         0.06         628.40         
cpu      128x128          8        11.08        0.50         722.34         
cpu      128x128          32       41.02        2.00         780.19         
cpu      256x256          1        5.88         0.25         170.00         
cpu      256x256          8        41.71        2.00         191.80         
cpu      256x256          32       156.74       8.00         204.16         
cpu      512x512          1        21.88        1.00         45.71          
cpu      512x512          8        154.99       8.00         51.62          
cpu      512x512          32       636.72       32.00        50.26          
cuda     64x64           1        0.44         8.31         2281.16        
cuda     64x64           8        0.41         9.52         19507.32       
cuda     64x64           32       0.42         13.65        76762.47       
cuda     128x128          1        0.41         9.28         2457.38        
cuda     128x128          8        0.40         13.67        20042.33       
cuda     128x128          32       0.40         30.17        80087.14       
cuda     256x256          1        0.39         12.67        2547.11        
cuda     256x256          8        0.40         30.18        20145.69       
cuda     256x256          32       0.75         96.19        42763.00       
cuda     512x512          1        0.39         26.19        2593.10        
cuda     512x512          8        0.76         96.20        10571.72       
cuda     512x512          32       2.25         360.21       14237.00       

================================================================================
Performance Analysis
================================================================================
Best GPU performance: 80087.14 images/sec (128x128, batch=32)
Best CPU performance: 2892.69 images/sec (64x64, batch=32)
GPU speedup: 27.69x

Benchmark completed!
```

Interesting points:

- On the CPU this PR makes the performance a LOT worse, more than 10x
- On the GPU the performance is pretty much unchanged, memory requirement goes even down a bit


Definitely want to follow up a bit on the CPU performance before merging, thre might be some interesting lessons here.

PS: God, I love uv